[benchmarks] Remove some duplication
