+++ diesel_bench/benches/diesel_benches.rs
@@ -1,3 +1,7 @@
+#[path = "common.rs"]
+mod common;
+
+use common::*;
 use super::Bencher;
 use diesel::*;
 
@@ -9,45 +13,6 @@ type TestConnection = MysqlConnection;
 
 #[cfg(feature = "sqlite")]
 type TestConnection = SqliteConnection;
-
-table! {
-    users {
-        id -> Integer,
-        name -> Text,
-        hair_color -> Nullable<Text>,
-    }
-}
-
-table! {
-    posts {
-        id -> Integer,
-        user_id -> Integer,
-        title -> Text,
-        body -> Nullable<Text>,
-    }
-}
-
-table! {
-    comments {
-        id -> Integer,
-        post_id -> Integer,
-        text -> Text,
-    }
-}
-
-joinable!(comments -> posts (post_id));
-joinable!(posts -> users (user_id));
-allow_tables_to_appear_in_same_query!(users, posts, comments);
-
-#[derive(
-    PartialEq, Eq, Debug, Clone, Queryable, Identifiable, Insertable, AsChangeset, QueryableByName,
-)]
-#[diesel(table_name = users)]
-pub struct User {
-    pub id: i32,
-    pub name: String,
-    pub hair_color: Option<String>,
-}
 
 #[derive(Debug, PartialEq, Eq, Queryable, Clone, Insertable, AsChangeset)]
 #[diesel(table_name = users)]
@@ -66,16 +31,6 @@ impl NewUser {
     }
 }
 
-#[derive(PartialEq, Eq, Debug, Clone, Queryable, Identifiable, Associations, QueryableByName)]
-#[diesel(belongs_to(User))]
-#[diesel(table_name = posts)]
-pub struct Post {
-    pub id: i32,
-    pub user_id: i32,
-    pub title: String,
-    pub body: Option<String>,
-}
-
 #[derive(Insertable)]
 #[diesel(table_name = posts)]
 pub struct NewPost {
@@ -92,14 +47,6 @@ impl NewPost {
             body: body.map(|b| b.into()),
         }
     }
-}
-
-#[derive(PartialEq, Eq, Debug, Clone, Queryable, Identifiable, Associations)]
-#[diesel(belongs_to(Post))]
-pub struct Comment {
-    id: i32,
-    post_id: i32,
-    text: String,
 }
 
 #[derive(Debug, Clone, Copy, Insertable)]
